add unit tests for refresh grant user validation
